pcap-util: clean up the byte-swapping macros.
Use the same code style as PCAP_BSWAP_64() for PCAP_BSWAP_32() and
PCAP_BSWAP_16().

For all of them:

- use uintNN_t types rather than u_short or u_int;
- use UINTnn_C macros for the byte masks, so that the appropriate suffix
  is added (for example, if uint64_t is unsigned long, it will append UL
  rather than ULL to the constant if necessary);
- put the (uintNN_t)(y) in parentheses;
- put a final cast to the desired type around the entire expression,
  just as we were already doing for PCAP_BSWAP_16().

Expand the comment before the macros to give additional information.

Remove a no-longer-applies comment and and a now-unnecessary cast from
testprogs/translatetest.c.

+++ testprogs/translatetest.c
@@ -286,12 +286,7 @@ test_PCAP_BSWAP_64(const char *arg)
 	snprintf(buf, sizeof(buf), "0x%016" PRIx64, i);
 	if (pcapint_strcasecmp(buf, arg))
 		goto fail;
-	/*
-	 * The current definition of PCAP_BSWAP_64() is an unsigned long long,
-	 * which does not match PRIx64 on 64-bit architectures, so cast it to
-	 * an uint64_t, which does match, to avoid a -Wformat.
-	 */
-	printf("OK: 0x%016" PRIx64 "\n", (uint64_t)PCAP_BSWAP_64(i));
+	printf("OK: 0x%016" PRIx64 "\n", PCAP_BSWAP_64(i));
 	return EX_OK;
 fail:
 	fprintf(stderr, "ERROR: failed parsing \"%s\"\n", arg);
